# Backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Fast Accounts API
  version: 0.1.0
  description: |
    Domain-grouped HTTP API for AI-Accounts (Fast Accounts parity). JSON bodies use **camelCase** keys;
    URL path segments use **kebab-case**. Aligns with Backend/.cursorrules_backend.
...
paths:
...
  /api/v1/auth/sign-up:
    post:
      tags: [Auth]
      summary: Register user and company (email OTP required before login)
      operationId: authSignUp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignUpRequest"
      responses:
        "201":
          description: Pending email verification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignUpPendingResponse"
  /api/v1/auth/verify-email:
    post:
      tags: [Auth]
      summary: Verify signup email with OTP and receive tokens
      operationId: authVerifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokensResponse"
  /api/v1/auth/resend-otp:
    post:
      tags: [Auth]
      summary: Resend signup or password-reset OTP
      operationId: authResendOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResendOtpRequest"
      responses:
        "200":
          description: OTP reissued or generic acknowledgement
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/SignUpPendingResponse"
                  - $ref: "#/components/schemas/AuthMessageResponse"
  /api/v1/auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request password reset OTP
      operationId: authForgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "200":
          description: Always returns generic message (no email enumeration)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMessageResponse"
  /api/v1/auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password using email OTP
      operationId: authResetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMessageResponse"
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Password login (requires verified email)
      operationId: authLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokensResponse"
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token (rotates refresh token)
      operationId: authRefresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokensResponse"
...
  /api/v1/companies:
    get:
      tags: [Companies]
      summary: List companies for current user
      operationId: listCompanies
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
    post:
      tags: [Companies]
      summary: Create additional company for current user
      operationId: createCompany
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompanyRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  companyId:
                    type: string
                  name:
                    type: string
  /api/v1/companies/{companyId}/switch:
    post:
      tags: [Companies]
      summary: Set active company context (returns new token pair)
      operationId: switchCompany
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/companyId"
      responses:
        "200":
          description: New tokens with selected companyId claim
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokensResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    companyId:
      name: companyId
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    SignUpRequest:
      type: object
      required: [email, password, firstName, lastName, companyName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        firstName:
          type: string
        lastName:
          type: string
        companyName:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    SignUpPendingResponse:
      type: object
      properties:
        message:
          type: string
        email:
          type: string
          format: email
        expiresAt:
          type: string
          format: date-time
        devOtp:
          type: string
          nullable: true
          description: Only when AUTH_EXPOSE_OTP_IN_RESPONSE=true (local dev)
    VerifyEmailRequest:
      type: object
      required: [email, otpCode]
      properties:
        email:
          type: string
          format: email
        otpCode:
          type: string
          pattern: "^\\d{6}$"
    ResendOtpRequest:
      type: object
      required: [email, purpose]
      properties:
        email:
          type: string
          format: email
        purpose:
          type: string
          enum: [signup, password_reset]
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    ResetPasswordRequest:
      type: object
      required: [email, otpCode, newPassword]
      properties:
        email:
          type: string
          format: email
        otpCode:
          type: string
          pattern: "^\\d{6}$"
        newPassword:
          type: string
          minLength: 8
    CreateCompanyRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 200
...
    AuthMessageResponse:
      type: object
      properties:
        message:
          type: string
...
    AuthTokensResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          example: bearer
...